Implemented double chess

+++ BattleChess3.UI/ViewModel/MapsViewModel.cs
@@ -1,8 +1,8 @@
-﻿using BattleChess3.Core.Model;
+﻿using System.Collections.ObjectModel;
+using BattleChess3.Core.Model;
 using BattleChess3.Core.Model.Figures;
 using BattleChess3.UI.Services;
 using GalaSoft.MvvmLight;
-using System.Collections.ObjectModel;
 
 namespace BattleChess3.UI.ViewModel;
 
@@ -11,27 +11,9 @@ public sealed class MapsViewModel : ViewModelBase, IDisposable
     private readonly IMapService _mapService;
     private readonly IPlayerService _playerService;
 
-    private MapBlueprint _selectedMap = MapBlueprint.None;
-    public MapBlueprint SelectedMap
-    {
-        get => _selectedMap;
-        set => Set(ref _selectedMap, value);
-    }
-
     private IList<MapBlueprint> _maps = Array.Empty<MapBlueprint>();
-    public IList<MapBlueprint> Maps
-    {
-        get => _maps;
-        private set
-        {
-            if (value.All(x => x.MapPath != _selectedMap.MapPath))
-            {
-                SelectedMap = value.FirstOrDefault()
-                    ?? MapBlueprint.None;
-            }
-            Set(ref _maps, value);
-        }
-    }
+
+    private MapBlueprint _selectedMap = MapBlueprint.None;
 
     public MapsViewModel(
         IMapService mapService,
@@ -44,6 +26,29 @@ public sealed class MapsViewModel : ViewModelBase, IDisposable
         _mapService.MapsChanged += OnMapsChanged;
     }
 
+    public MapBlueprint SelectedMap
+    {
+        get => _selectedMap;
+        set => Set(ref _selectedMap, value);
+    }
+
+    public IList<MapBlueprint> Maps
+    {
+        get => _maps;
+        private set
+        {
+            if (value.All(x => x.MapPath != _selectedMap.MapPath))
+                SelectedMap = value.FirstOrDefault()
+                              ?? MapBlueprint.None;
+            Set(ref _maps, value);
+        }
+    }
+
+    public void Dispose()
+    {
+        _mapService.MapsChanged -= OnMapsChanged;
+    }
+
     private void OnMapsChanged(object? sender, IList<MapBlueprint> maps)
     {
         Maps = new ObservableCollection<MapBlueprint>(maps);
@@ -51,10 +56,7 @@ public sealed class MapsViewModel : ViewModelBase, IDisposable
 
     internal void DeleteSelectedMap()
     {
-        if (SelectedMap != MapBlueprint.None)
-        {
-            _mapService.Delete(SelectedMap);
-        }
+        if (SelectedMap != MapBlueprint.None) _mapService.Delete(SelectedMap);
     }
 
     internal void SaveSelectedMap(string identifier, IEnumerable<ITile> board)
@@ -68,14 +70,9 @@ public sealed class MapsViewModel : ViewModelBase, IDisposable
             }).ToArray(),
             MapPath = $"Resources/Maps/{identifier}.map",
             PreviewPath = $"./Resources/Maps/{identifier}.png",
-            StartingPlayer = _playerService.CurrentPlayer.Id,
+            StartingPlayer = _playerService.CurrentPlayer.Id
         };
 
         _mapService.Save(map);
-    }
-
-    public void Dispose()
-    {
-        _mapService.MapsChanged -= OnMapsChanged;
     }
 }
